author/requirements.txt
openai==0.27.8
Flask==2.1.0
gunicorn==20.1.0
google-cloud-secret-manager==2.16.1
google-cloud-datastore==2.16.1
